chore(typescript): lint declaration files
