+++ package.json
@@ -18,16 +18,14 @@
     "publish:canary": "npm publish --tag canary",
     "lint": "standard",
     "lint:fix": "standard --fix",
+    "lint:types": "dtslint types ",
     "test:db:start": "docker-compose -f test/docker/docker-compose.yml up",
     "test:db:stop": "docker-compose -f test/docker/docker-compose.yml down"
   },
-  "types": "types/index.d.ts",
+  "types": "types",
   "files": [
     "dist",
-    "index.js",
-    "providers.js",
-    "adapters.js",
-    "client.js"
+    "types"
   ],
   "license": "ISC",
   "dependencies": {
@@ -48,12 +46,21 @@
     "@babel/cli": "^7.8.4",
     "@babel/core": "^7.9.6",
     "@babel/preset-env": "^7.9.6",
+    "@types/react": "^16.9.35",
+    "@types/react-dom": "^16.9.8",
+    "@types/webpack": "^4.41.17",
     "autoprefixer": "^9.7.6",
     "babel-preset-minify": "^0.5.1",
     "babel-preset-preact": "^2.0.0",
     "cssnano": "^4.1.10",
+    "dtslint": "^3.6.10",
+    "next": "^9.4.4",
     "postcss-cli": "^7.1.1",
     "postcss-nested": "^4.2.1",
-    "standard": "^14.3.3"
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "standard": "^14.3.3",
+    "tsd": "^0.11.0",
+    "typescript": "^3.9.5"
   }
 }
