+++ types/test.ts
@@ -1,0 +1,3 @@
+import NextAuth from 'next-auth';
+
+NextAuth({}, {}, {});
